backend/security_movements/serializers.py
from rest_framework import serializers
from security_movements.models import Movement

class MovementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movement
        fields = "__all__"