backend/requirements.txt
Django==3.2.14
djangorestframework==3.13.1
